+++ base/CMakeLists.txt
@@ -1,12 +1,14 @@
+# SPDX-License-Identifier: Apache-2.0
+
+set(DTC_OVERLAY_FILE dtc_shell.overlay)
+
+set(BOARD nrf52840dongle_nrf52840)
 
 cmake_minimum_required(VERSION 3.20.0)
 
-set(BOARD thingy52_nrf52832)
-set(CONF_FILE conf_file.conf)
 
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
-project(mesh)
-
+project(base)
 
 FILE(GLOB app_sources src/*.c)
 target_sources(app PRIVATE ${app_sources})
