+++ base/dtc_shell.overlay
@@ -1,6 +1,6 @@
 / {
 	chosen {
-		zephyr,console = &cdc_acm_uart0;
+		zephyr,shell-uart = &cdc_acm_uart0;
 	};
 };
