+++ base/flash.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+nrfutil pkg generate --hw-version 52 --sd-req=0x00 \
+	--application build/zephyr/zephyr.hex \
+	--application-version 1 base.zip
+
+sudo chmod a+rw /dev/ttyACM0
+
+nrfutil dfu usb-serial -pkg base.zip -p /dev/ttyACM0
