+++ base/prj.conf
@@ -1,0 +1,104 @@
+CONFIG_SYSTEM_WORKQUEUE_STACK_SIZE=2048
+
+CONFIG_BT_SETTINGS=n
+CONFIG_FLASH=y
+CONFIG_FLASH_MAP=y
+CONFIG_NVS=y
+CONFIG_SETTINGS=n
+CONFIG_HWINFO=y
+
+CONFIG_BT_OBSERVER=y
+CONFIG_BT_CTLR_DUP_FILTER_LEN=0
+CONFIG_BT_CTLR_LE_ENC=n
+CONFIG_BT_DATA_LEN_UPDATE=n
+CONFIG_BT_PHY_UPDATE=n
+CONFIG_BT_CTLR_CHAN_SEL_2=n
+CONFIG_BT_CTLR_MIN_USED_CHAN=n
+CONFIG_BT_CTLR_PRIVACY=n
+
+CONFIG_BT_MESH_RX_SEG_MSG_COUNT=100
+
+CONFIG_BT_PERIPHERAL=y
+
+CONFIG_BT=y
+CONFIG_BT_TINYCRYPT_ECC=y
+CONFIG_BT_L2CAP_TX_BUF_COUNT=5
+
+CONFIG_BT_MESH=y
+CONFIG_BT_MESH_MODEL_EXTENSIONS=y
+CONFIG_BT_MESH_RELAY=y
+CONFIG_BT_MESH_FRIEND=y
+CONFIG_BT_MESH_PB_GATT=y
+CONFIG_BT_MESH_PB_ADV=y
+CONFIG_BT_MESH_GATT_PROXY=y
+
+#CONFIG_BT_MESH_LOW_POWER=y
+#CONFIG_BT_MESH_LPN_SCAN_LATENCY=30
+#CONFIG_BT_MESH_LPN_RECV_DELAY=40
+#CONFIG_BT_MESH_LPN_POLL_TIMEOUT=300
+
+CONFIG_BT_MESH_SUBNET_COUNT=2
+CONFIG_BT_MESH_APP_KEY_COUNT=2
+CONFIG_BT_MESH_MODEL_GROUP_COUNT=2
+CONFIG_BT_MESH_LABEL_COUNT=3
+
+CONFIG_BT_DEBUG_LOG=n
+
+CONFIG_BT_MESH_DEBUG=y
+CONFIG_BT_MESH_DEBUG_MODEL=y
+
+#------------------------------ENABLE USB---------------------------------------
+#Add this line if building for the Arduino Sense Board
+#CONFIG_BOARD_ARDUINO_NANO_33_BLE_EN_USB_CONSOLE=y
+CONFIG_USB_DEVICE_STACK=y
+CONFIG_USB_DEVICE_PRODUCT="USB Console Tute"
+
+CONFIG_SERIAL=y
+CONFIG_CONSOLE=y
+CONFIG_UART_CONSOLE=y
+CONFIG_UART_LINE_CTRL=y
+#------------------------------------------------------------------------------
+
+CONFIG_CBPRINTF_FP_SUPPORT=y
+
+#-----------------------------SHELL_CONF--------------------------------------
+CONFIG_SHELL_BACKEND_SERIAL_CHECK_DTR=y
+CONFIG_SHELL_BACKEND_SERIAL_INIT_PRIORITY=51
+CONFIG_SHELL=y
+CONFIG_SHELL_BACKENDS=y
+CONFIG_SHELL_BACKEND_SERIAL=y
+
+CONFIG_SHELL_PROMPT_UART="base:~$"
+
+CONFIG_SHELL_CMDS=y
+CONFIG_SHELL_TAB=y
+CONFIG_SHELL_TAB_AUTOCOMPLETION=y
+CONFIG_SHELL_VT100_COLORS=y
+CONFIG_KERNEL_SHELL=y
+#-----------------------------------------------------------------------------
+
+#-----------------------------SHELL_LOGGING-----------------------------------
+CONFIG_LOG=y
+CONFIG_LOG_PRINTK=y
+#-----------------------------------------------------------------------------
+CONFIG_USB_DEVICE_VID=0xC553
+CONFIG_USB_DEVICE_PID=0x4011
+CONFIG_USB_DEVICE_MANUFACTURER="Hugh R"
+
+
+
+################# PROVISIONING #########################
+CONFIG_BT_MESH_PROVISIONER=y
+CONFIG_BT_MESH_PROV_DEVICE=n
+CONFIG_BT_MESH_CDB=y
+CONFIG_BT_MESH_CDB_NODE_COUNT=16
+CONFIG_BT_MESH_CDB_SUBNET_COUNT=3
+CONFIG_BT_MESH_CDB_APP_KEY_COUNT=3
+
+CONFIG_BT_MESH_CFG_CLI=y
+CONFIG_BT_MESH_HEALTH_CLI=y
+
+CONFIG_MAIN_STACK_SIZE=1408
+CONFIG_MAIN_THREAD_PRIORITY=-2
+
+CONFIG_BT_MESH_RELAY_RETRANSMIT_COUNT=3
