+++ base/src/board.h
@@ -6,10 +6,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-void board_output_number(bt_mesh_output_action_t action, uint32_t number);
-
-void board_prov_complete(void);
-
-int board_init(struct k_work *button_work);
+int board_init(void);
 
 void board_led_set(bool val);
