--- node_base/CMakeLists.txt
@@ -1,7 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-set(BOARD nrf52840dongle_nrf52840)
-cmake_minimum_required(VERSION 3.20.0)
-find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
-project(blinky)
-
-target_sources(app PRIVATE src/main.c)
