--- node_base/prj.conf
@@ -1,86 +1,0 @@
-CONFIG_GPIO=y
-#Enable this for Arduino Nano board
-#CONFIG_BOARD_ARDUINO_NANO_33_BLE_EN_USB_CONSOLE=y
-
-#-----------------------------------------------------------------------------
-CONFIG_USB_DEVICE_STACK=y
-CONFIG_USB_DEVICE_PRODUCT="Zephyr USB console sample"
-
-CONFIG_SERIAL=y
-CONFIG_CONSOLE=y
-CONFIG_UART_CONSOLE=y
-CONFIG_UART_LINE_CTRL=y
-#-----------------------------------------------------------------------------
-
-#--------------------------------USB OPTIONS----------------------------------
-CONFIG_USB_DEVICE_PRODUCT="Particle Argon - Zephyr"
-CONFIG_USB_DEVICE_MANUFACTURER="Wilfred MK"
-CONFIG_USB_DEVICE_VID=0xC553
-CONFIG_USB_DEVICE_PID=0x4011
-#-----------------------------------------------------------------------------
-
-CONFIG_BT=y
-CONFIG_BT_BROADCASTER=y
-CONFIG_BT_OBSERVER=y
-#CONFIG_BT_DEBUG_LOG=y
-CONFIG_BT_DEVICE_NAME="base"
-
-
-
-
-
-
-#CONFIG_INIT_STACKS=y
-CONFIG_MAIN_STACK_SIZE=1408
-CONFIG_SYSTEM_WORKQUEUE_STACK_SIZE=2048
-# The Bluetooth API should not be used from a preemptive thread:
-CONFIG_MAIN_THREAD_PRIORITY=-2
-
-CONFIG_BT=y
-CONFIG_BT_TINYCRYPT_ECC=y
-#CONFIG_BT_DEBUG_LOG=y
-CONFIG_BT_OBSERVER=y
-CONFIG_BT_BROADCASTER=y
-
-CONFIG_BT_MESH=y
-CONFIG_BT_MESH_SUBNET_COUNT=1
-CONFIG_BT_MESH_APP_KEY_COUNT=1
-CONFIG_BT_MESH_ADV_BUF_COUNT=10
-CONFIG_BT_MESH_TX_SEG_MSG_COUNT=3
-CONFIG_BT_MESH_RX_SEG_MAX=32
-CONFIG_BT_MESH_MODEL_GROUP_COUNT=2
-CONFIG_BT_MESH_LABEL_COUNT=0
-CONFIG_BT_MESH_CFG_CLI=y
-CONFIG_BT_MESH_HEALTH_CLI=y
-CONFIG_BT_MESH_BEACON_ENABLED=n
-CONFIG_BT_MESH_RELAY=y
-CONFIG_BT_MESH_RELAY_RETRANSMIT_COUNT=3
-
-CONFIG_BT_MESH_PROVISIONER=y
-CONFIG_BT_MESH_PROV_DEVICE=n
-CONFIG_BT_MESH_CDB=y
-CONFIG_BT_MESH_CDB_NODE_COUNT=16
-CONFIG_BT_MESH_CDB_SUBNET_COUNT=3
-CONFIG_BT_MESH_CDB_APP_KEY_COUNT=3
-
-CONFIG_BT_SETTINGS=y
-CONFIG_FLASH=y
-CONFIG_FLASH_PAGE_LAYOUT=y
-CONFIG_FLASH_MAP=y
-CONFIG_NVS=y
-CONFIG_SETTINGS=y
-CONFIG_BT_MESH_RPL_STORE_TIMEOUT=600
-
-#CONFIG_BT_MESH_DEBUG=y
-#CONFIG_BT_MESH_DEBUG_SETTINGS=y
-#CONFIG_BT_MESH_DEBUG_NET=y
-#CONFIG_BT_MESH_DEBUG_MODEL=y
-#CONFIG_BT_MESH_DEBUG_TRANS=y
-#CONFIG_BT_MESH_DEBUG_ACCESS=y
-
-#CONFIG_BT_MESH_LOW_POWER=y
-#CONFIG_BT_MESH_DEBUG_PROV=y
-#CONFIG_BT_MESH_DEBUG_LOW_POWER=y
-#CONFIG_BT_MESH_DEBUG_BEACON=y
-#CONFIG_BT_MESH_DEBUG_CRYPTO=y
-#CONFIG_BT_MESH_DEBUG_ADV=y
